More type fixes. We now build with -Wconversion, on this platform (64-bit Linux) at least.

+++ main/PreferencesDialog.h
@@ -90,7 +90,7 @@ protected:
     int m_spectrogramSmoothing;
     int m_spectrogramXSmoothing;
     int m_propertyLayout;
-    float m_tuningFrequency;
+    double m_tuningFrequency;
     int m_audioDevice;
     int m_resampleQuality;
     bool m_resampleOnLoad;
